Rename main modal entry to ModalDashboard
